refa: break train.py in two files
